Moved notification struct to pkg so it can be linked by other software
