+++ src/apps/chifra/internal/scrape/scrape_consolidate.go
@@ -14,6 +14,7 @@ import (
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/file"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/index"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/logger"
+	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/notify"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/sigintTrap"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/utils"
@@ -117,8 +118,8 @@ func (bm *BlazeManager) Consolidate(blocks []base.Blknum) (error, bool) {
 				report.FileSize = file.FileSize(chunkPath)
 				report.Report()
 			}
-			if err := Notify(Notification[string]{
-				Msg:     MessageChunkWritten,
+			if err := Notify(notify.Notification[string]{
+				Msg:     notify.MessageChunkWritten,
 				Meta:    bm.meta,
 				Payload: chunkRange.String(),
 			}); err != nil {
@@ -163,8 +164,8 @@ func (bm *BlazeManager) Consolidate(blocks []base.Blknum) (error, bool) {
 	nAppsNow := int(file.FileSize(stageFn) / asciiAppearanceSize)
 	bm.report(len(blocks), int(bm.PerChunk()), nChunks, nAppsNow, nAppsFound, nAddrsFound)
 
-	if err := Notify(Notification[string]{
-		Msg:     MessageStageUpdated,
+	if err := Notify(notify.Notification[string]{
+		Msg:     notify.MessageStageUpdated,
 		Meta:    bm.meta,
 		Payload: newRange.String(),
 	}); err != nil {
